requirements.txt
 crewai[tools]
langchain-openai
python-dotenv
Flask
gunicorn
google-cloud-secret-manager

